Process lines with '=>' only with find_definitions
Previously, the conditional that processes procedure renamings inside
a type definition was run even if only the current scope was requested
(fortran-tags.py -s). This lead to an elisp error in certain
situations, but was unnoticed because it didn't seem to affect any
functionality. This is now fixed: the conditional is run only if
find_definitions = True, i.e., while we are building the TAGS file.