Add function to initialize global Elisp variables
The first time Fortran-tags is called, things like the integrity of
the tags file and the remote protocol, if present, need to be checked
and certain global variables need to be initialized
accordingly. Perform these checks in a separate function that can be
called from multiple places depending on which function the user
invokes first.